gatchat: fix unref
If the new_bytes handler processes a modem shutdown URC, it results in
a call to the g_at_chat_unref function, which includes a call to the
chat_cleanup function, which clears the list of URC handlers that
continue to be traversed after the current handler has completed, and
may lead to errors. To correct the situation, the at_chat_ref function
has been added, which is called at the beginning of the new_bytes
handler. Added a call to the at_chat_unref function at the end of the
new_bytes handler. Added at_chat_ref function call to g_at_chat_clone.
The previous partial solution to this problem has been removed.