tools: Add provision.py
Introduce a new tool that will convert the intermediate JSON format
(documented in doc/provision.rst) to the binary provisioning format,
which will be used by ofonod's default provisioning plugin for
automatically setting up carrier specific settings.

The tool also supports import of and conversion of
'mobile-broadband-provider-info' XML files to the new intermediate JSON
file format.  This is accomplished using:
  % tools/provisiontool mbpi-convert --outfile=provision.json

Conversion of JSON intermediate format to binary format is accomplished
using:
  % tools/provisiontool generate --infile=provision.json

By default, the output will be placed in the same directory in the file
'provision.db'.  Alternatively, the output file can be specified using
the --outfile option.

Finally, the tool supports a simple selftest method, which can be
invoked as follows:
  % tools/provisiontool selftest