stkagent: Use l_utf8_get_codepoint
The GetKey() validation path ensures that the agent returns a single
UTF8 character.  It performs g_utf8_strlen with a maximum size of 10
bytes in order to minimize unnecessary processing.  Convert this code
to use the more fitting l_utf8_get_codepoint function, which will read
a single UTF8 character from the stream.