qmi: Introduce qmi_device_new_qmux
In preparation for QRTR support, remove the existing qmi_device_new API
and rename it to qmi_device_new_qmux.  Since QMUX usually runs on top of
a character device, have the constructor take in the character device
path, instead of having the caller open it manually.

As part of this refactoring, introduce the initial qmi_device_ops
structure which will be used to abstract qmi operation over QMUX and
QRTR.