qmi: Introduce shutdown operation
On QMUX the shutdown operation waits for all low-level service handles
to be released, then calls the shutdown callback / destroy functions.
Low level service handles on QMUX protocol are released via
qmi_service_unref, which in turn sends a QMUX control message.

On the other hand, in QRTR, there is no need to explicitly release
handles.  They are released automatically when the underlying QRTR
socket is closed.

Abstract the shutdown operation behind a shutdown method in the ops
structure, and move the QMUX specific implementation there.  While here,
change the return signature of qmi_device_shutdown to return an int,
such that better error reporting can be supported.  For example, QRTR
might choose not to implement the shutdown method entirely, in which
case an -EOPNOTSUP will be returned.