qmi: Discover timeout could cause a crash
The DISCOVERY_DONE macro assumes that the provided data is valid.
Unofortunately if the queue is empty due to a discover timeout, the
value would either be NULL or the next "discovery" item on the queue
which is not necessarily a discover action. This change addresses
each of these possibilities.