qmi: Initial unit test POC for QRTR
Looking for some feedback on my approach here. I need
to be able to fake the AF_QIPCRTR socket, but still
utilize the socket APIs so the tests can inspect and
send sockaddr_qrtr structs.

The approach that I am taking involves
1. Injecting certain dependencies to a private function
   used only internally in qmi.c and in the unit tests
2. Utilizing the linker's wrap directive to allow
   sendto and recvfrom to be overridden.

This allows me to inject a UNIX domain socket while
still validing the correct usage of sendto and
recvfrom.

I am open to trying alternative approaches if there
is a different strategy you would like to see.