qmi: Fix printing the wrong type for services
The __debug_msg refactoring introduced a bug where the service
transactions would not display the correct type in the log. The service
transaction type is shifted to the left by one bit compared to the
control transaction type so we need to perform a shift to be able to
share a common debug logging function.

Fixes: afa96ae4fa99 ("qmi: Clean up the __debug_msg function")