qmi: Initialize group_id for control requests
valgrind found that for control requests, the group_id was not being
set. The requests are being allocated using l_malloc which does not
initialize memory to 0, unlike l_new. The impact is that cleanup of
the requests during service closure or modem shutdown might
incorrectly remove a control request from the queue. The fix is to
always initialize this field to 0.

Fixes: 0a4591e439ba ("qmi: Add an abstract group id to services and requests")