udevng: Detect embedded qmi qrtr modems
Embedded qmi qrtr modems are identified by the existence of
rmnet_ipaX and rmnet_dataX devices. Add a new "embedded" modem type
so that these devices can be collected during enumeration and then
configured for use by the gobi plugin. Modems of this type will be
exposed as /gobiqrtr_X.

Retrieving the endpoint ID is the main complication. It requires
using an interface IOCTL and extended structure that is declared
in msm_rmnet.h. This header is available many different places on
the internet including in Android sources, but it is not available
in popular Linux distributions. The minimum set of declarations is
provided here so that the IOCTL may be used.