ofono: Add support for ofono main.conf settings
It would be useful to support some oFono wide configuration settings
that can be configured for the system.  Introduce a new
__ofono_get_config() function that will obtain the parsed settings file
as a pointer to l_settings.  The settings will be parsed from the
configuration directory set using CONFIGURATION_DIRECTORY environment
variable, or the default CONFIGDIR variable set during
configuration/compilation.