qmimodem: Remove unregister/cancel_all functions
gobi was the final user of these functions in its disable function. It
would perform this cleanup before sending a final message. This is not
necessary because gobi does not register for any notifications and the
only sends that might be in progress should only occur during
initialization. If a client actually needs this functionality it should
free the service and create a new one.

These functions are still used internally in qmi.