qmi: gprs: use Extended Data Bearer Technology
This TLV is reported by WDS "Event Report" indication and contains a
better representation of the current bearer compared to the Data Service
Capability TLV reported in the NAS Serving System indication.

 TLV:
   type       = "Extended Data Bearer Technology" (0x2a)
   length     = 16
   value      = 00:00:00:00:03:00:00:00:00:10:00:00:00:00:00:00
   translated = [ data_bearer_technology = '3gpp'
 		  radio_access_technology = '3gpp-lte'
		  extended_data_bearer_technology_3gpp = 'lte-fdd'
		  ...]

 TLV:
   type       = "Data Service Capability" (0x11)
   length     = 2
   value      = 01:0B
   translated = { [0] = 'lte '}

Some of the 5G and more esoteric technologies are not yet handled in
this commit.  Support for these technologies needs to be added in the
core first.

Modify the logic in the gprs driver to use this new mechanism.