qmi: combine qmux_client_create_data into qmi_request
Similarly to the previous commit, do not allocate a separate data
structure to hold client creation details, and instead allocate it as
a qmi_request specialization.  Since the lifetime of struct
qmux_client_create_data was already tied to the lifetime of the request
(and vice-versa), this change allows the extra allocation to be
optimized away.

While here, remove all vestiges of discovery and discovery_queue.  This
functionality is no longer needed since any data associated with the
request will be freed when struct qmi_request is freed.