gobi: Bring down the main interface at startup
The current gobi logic tried to set the qmi_wwan 'raw_ip' state based on
the results of wda_get_data_format.  If link_layer protocol was
'raw_ip', then 'raw_ip' setting on qmi_wwan was set to 'Y'.  For 802.3,
'raw_ip' setting was set to 'N'.

Unfortunately this was broken at some point and the setting could not be
toggled when the main network device was IFF_UP (default when USB device
is detected by the kernel).

Fix this by having .enable callback bring the device down at
initialization time.  oFono core will set the device IFF_UP once the
context is activated, and bring it down when the context is deactivated.