rmnet: Implement getting new interfaces
Add method to create N rmnet interfaces for use by the gobi driver.  The
rmnet interfaces are created in sequence, which should result in
sequential mux ids and interface names to be created (assuming no other
process creates rmnet interfaces in the background).

For now we hardcode the use of QMAPv5 as the aggregation protocol using
EGRESS_MAP_CHKSUMV5 and INGRESS_MAP_CHKSUMV5.  Support for QMAPv4 or
QMAP is also possible in the future, but no devices are currently
available for testing.