gobi: Add support for multiple active contexts
Add support for multiple contexts based on qmi_wwan 'passthrough' +
rmnet combination.  This is a newer and more flexible approach compared
to the legacy qmi_wwan 'add_mux' / 'del_mux' design.  Multiple contexts
are enabled as follows:
  - If the device supports passthrough, attempt to set the modem to use
    QMAPv5 for both downlink and uplink aggregation.
  - If successful, request rmnet module to create multiple rmnet
    network interfaces to use for each context / PDN that will be
    simultaneously active
  - Request WDS service handles for each rmnet device created

With the above setup, the main network interface MTU is changed to
16384, which is the maximum supported by the current qmi_wwan driver.
Note that modems might be capable of larger aggregation sizes,
particularly for the download direction.  However, such sizes will
exceed the maximum URB (corresponding to the MTU) which qmi_wwan
supports.