data: Add initial attach type to Verizon 'vzwinternet' context.
Verizon has a number of different Cellular access point names (APNs)
and contexts, depending on the application, per their documentation
here:

    https://www.verizon.com/support/knowledge-base-72601/

Some of those contexts must be used for initial attach (IA). For Voice
over LTE (VoLTE) applications, the 'vzwims' APN is such an IA context
and is currently marked as such.

However, for Internet of Things (IoT) or machine-to-machine (M2M)
applications, the 'vzwinternet' APN is also such an IA context but is
not presently marked as such. This change adds the 'ia' type to this
APN context.