data: Add tags to Verizon APN contexts.
Verizon has a number of different Cellular access point names (APNs)
and contexts, depending on the application, per their documentation
here:

    https://www.verizon.com/support/knowledge-base-72601/

One partition of those contexts are appropriate for LTE and/or Voice
over LTE (VoLTE) applications while another is appropriate for
Internet of Things (IoT) or machine-to-machine (M2M) applications.

This applies the 'iot', 'm2m', and 'volte' tags to the three current
Verizon APN contexts such that no tags or some combination of those
tags produce the desired set of APN contexts for the desired
application.