Patch for crash parsing malformed MBIM packets
Hi oFono devs,

I have a client using oFono in the stack for an IoT gateway with a Cinterion PLS62-W
modem. On a small number of units, at random, this modem sends a specific URC as an MBIM
packet.  oFono parses the ASCII bytes as an MBIM header, leading to a wild pointer and a
segfault:
    MBIM:n_iov: 3, 60
    MBIM:> 03 00 00 00 3c 00 00 00 0b 00 00 00 01 00 00 00 ....<...........
    MBIM:  00 00 00 00 53 3f be eb 14 fe 44 67 9f 90 33 a2 ....S?....Dg..3.
    MBIM:  23 e5 6c 3f 02 00 00 00 00 00 00 00 0c 00 00 00 #.l?............
    MBIM:  00 00 00 00 00 00 00 00 00 00 00 00 ............
    MBIM:< 2b 43 4d 47 4c 3a 20 31 2c 31 2c 2c +CMGL: 1,1,,
    MBIM:hdr->len: 824195660, header_size: 12, header_offset: 12
    MBIM:segment_bytes_remaining: 824195648
    MBIM:hdr->len: 824195660, header_size: 12, header_offset: 12
    MBIM:segment_bytes_remaining: 824195309
    MBIM:hdr->len: 824195660, header_size: 12, header_offset: 12
    MBIM:segment_bytes_remaining: 824195058

I added a safety check and after a few days of testing it seems to have solved the
problem. I've attached a patch against the current master in hopes it'll be useful.

Thanks,

Morgan Hughes