mbim: Add support for nested fixed arrays in arrays/structs
In some arrays, a nested array can be a fixed array coupled with an
OL-pair, which breaks the existing assumptions of arrays and structs,
leading to wrong data being parsed. Or the parent array can have no
OL-pair in this condition.

For example:
The UICC low-level access data has a struct array, but the array itself
has no OL-pair, instead the OL-pair is of the struct, so the position is
no longer incremented in the array parser so the nested struct can parse
its OL-pair properly. Instead the global iterator position is increased,
so the properties after the OL-pair can be parsed properly.

The method for detecting a fixed array is also flawed, because a nested
array can be fixed, e.g. in the example above, the AID array inside the
struct array is a fixed array, but parsing fails due to it being treated
as a variable array.

Introduce a new argument type to parse fixed nested arrays in order to
solve this. Also remove out_n_elem for arrays, since it can be a
separate field away from the array itself.